Simple Cat Toy: Ball of Yarn
Blanchard's rubric for AI students wants the game to demonstrate physics of some kind, so I decided to add a simple toy that bounces and is affected by collisions with the cat + the walls. It has a "trail" effect that makes it look kind of like a ball of yarn. I added a "playing" state to cat.cs that makes the cat move to the position of the ball of yarn.